\name{freegroup-package}
\alias{freegroup-package}
\alias{freegroup}
\docType{package}
\title{
\packageTitle{freegroup}
}
\description{
\packageDescription{freegroup}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{freegroup}
\packageIndices{freegroup}

}
\author{
\packageAuthor{freegroup}

Maintainer: \packageMaintainer{freegroup}
}
\keyword{ package }
\examples{
u <- rfree(10,6,2)
u^as.free('x')
abelianize(u)

}
