\name{outer}
\alias{outer}
\alias{outerperm_single_X}
\alias{outerperm_single_f}
\alias{outerperm_vec}
\alias{outerperm}
\alias{outerinv}
\alias{outersubs_lowlevel}
\alias{outersubs}
\title{Outer automorphisms of the free group}
\description{Vectorized functionality to implement outer automorphisms of the free group}
\usage{
outerperm_single_X(X,f)
outerperm_single_f(X,f)
outerperm_vec(X,f)
outerperm(X,f)
outerinv(X,s)
outersubs_lowlevel(M,e,S)
outersubs(X,e,S)
}
\arguments{
  \item{X,S}{Object of class \code{free}}
  \item{f}{Permutation function}
  \item{M}{Single free group element, in two-row matrix form}
  \item{e}{Element to substitute; see details}
}
\details{

sdfasf 
  }
\author{Robin K. S. Hankin}


\note{
Note here
}
\examples{


}
